+++ packages/components-extra/src/components/Footer/components/FooterContainer.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const FooterContainer = styled.footer`
+  width: 100%;
+  ${({ theme: { palette } }): string => `
+    background-color: ${palette.primary.main};
+  `};
+`
+
+export default FooterContainer