+++ packages/components-extra/src/components/Footer/components/TopContainer.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+import { CardMedia } from '@material-ui/core'
+
+const TopContainer = styled(CardMedia)`
+  padding: 36px;
+`
+
+export default TopContainer